backend/main/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
channels==4.2.2
charset-normalizer==3.4.2
distlib==0.3.8
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.13.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.0
pillow==11.2.1
platformdirs==4.2.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.1.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
virtualenv==20.28.0
